+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 22.0.0
+
+##### Breaking
+- Updates the native Android bridge [from Braze Android SDK 42.3.1 to 43.0.0](https://github.com/braze-inc/braze-android-sdk/compare/v42.3.1...v43.0.0#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4ed).
+- Updates the native iOS bridge [from Braze Swift SDK 17.0.0 to 18.0.0](https://github.com/braze-inc/braze-swift-sdk/compare/17.0.0...18.0.0#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4ed).
+
 ## 21.0.0
 
 ##### Breaking